// src/api/bookmark-service/models/api-response-boolean.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface ApiResponseBoolean
 */
export interface ApiResponseBoolean {
    /**
     * 
     * @type {boolean}
     * @memberof ApiResponseBoolean
     */
    'data'?: boolean;
    /**
     * 
     * @type {string}
     * @memberof ApiResponseBoolean
     */
    'message'?: string;
    /**
     * 
     * @type {number}
     * @memberof ApiResponseBoolean
     */
    'status'?: number;
}