// src/api/bookmark-service/models/api-response-list-bookmark-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { BookmarkResponse } from './bookmark-response';

/**
 * 
 * @export
 * @interface ApiResponseListBookmarkResponse
 */
export interface ApiResponseListBookmarkResponse {
    /**
     * 
     * @type {Array<BookmarkResponse>}
     * @memberof ApiResponseListBookmarkResponse
     */
    'data'?: Array<BookmarkResponse>;
    /**
     * 
     * @type {string}
     * @memberof ApiResponseListBookmarkResponse
     */
    'message'?: string;
    /**
     * 
     * @type {number}
     * @memberof ApiResponseListBookmarkResponse
     */
    'status'?: number;
}